requirements.txt
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==28.1.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
hyperlink==21.0.0
icalendar==5.0.13
ics==0.7.2
idna==3.8
incremental==24.7.2
inflection==0.5.1
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
TatSu==5.8.3
tomli==2.0.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
zope.interface==7.0.3
